# Avoid creating context for eval if eval is not present